+++ R/nec.R
@@ -91,6 +91,8 @@ nec <- function(object, posterior = FALSE, xform = identity,
 #' @export
 nec.bayesnecfit <- function(object, posterior = FALSE, xform = identity,
                             prob_vals = c(0.5, 0.025, 0.975), ...) {
+  check_component_arg(list(...), object)
+  check_nec_no_dpar(list(...))
   chk_logical(posterior)
   if(!inherits(xform, "function")){ 
     stop("xform must be a function.")} 
@@ -137,6 +139,8 @@ nec.bayesnecfit <- function(object, posterior = FALSE, xform = identity,
 #' @export
 nec.bayesmanecfit <- function(object, posterior = FALSE, xform = identity,
                               prob_vals = c(0.5, 0.025, 0.975), ...) {
+  check_component_arg(list(...), object)
+  check_nec_no_dpar(list(...))
   chk_logical(posterior)
   if (!inherits(xform, "function")) {
     stop("xform must be a function.")